third_party/WebKit/LayoutTests/http/tests/navigation/ping-cross-origin-from-https-expected.txt

CONSOLE ERROR: line 25: [blocked] The page at 'https://127.0.0.1:8443/navigation/resources/ping-cross-origin-from-https-target.html' was loaded over HTTPS, but ran insecure content from 'http://localhost:8000/navigation/resources/save-Ping.php?test=cross-origin-from-https': this content should also be loaded over HTTPS.

This page calls testRunner.notifyDone().
